+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # [develop](https://github.com/adhearsion/punchblock)
   * Feature: Output component now responds to a Stop command on Asterisk
   * Feature: Input component on Asterisk now responds to Stop command
+  * Feature: started/stopped-speaking events are now handled
 
 # [v0.10.0](https://github.com/adhearsion/punchblock/compare/v0.9.2...v0.10.0) - [2012-03-19](https://rubygems.org/gems/punchblock/versions/0.10.0)
   * Feature: app_swift is now supported on Asterisk with a media_engine type of :swift